wl12xx: replace dummy_join with ROC/CROC commands
With ROC/CROC, we no longer need the dummy_join.
We enter ROC:
1. when going out of idle state.
2. after disassoc
3. on scan complete (when not associated)

We stop ROC:
1. when entering idle state
2. after association (before join)
3. before starting a scan

The ROC command automatically opens the filters, so we no
longer need to configure them explicitly.

Replace the WL1271_FLAG_JOINED with a anew WL1271_FLAG_ROC,
indicating the fw was configured to ROC. Wrap the
wl1271_cmd_roc/croc functions in order to clear/set
this flag automatically.