wl12xx: don't use 11b rates in p2p (workaround)
When configuring basic rate set in p2p, filter out 11b rates.

11b rates shouldn't be used in p2p interfaces. otoh, they are
mandatory rates for 2GHZ band (which means we must be able to
receive packets in this frame, but shouldn't use it for tx).

Thus, we need to filter out the 11b rates for our rate policy
(used for tx only).