wl12xx: Allow disabling the bus FW logger
When possible, it is sometimes more convenient to retrieve logs by
connecting to the FW debug pins. This will not be possible if the FW
logger is configured to route logs to the host.

Allow disabling the bus FW logger via a debugfs entry. Please note that
this setting will only take effect after the FW reboots.

Signed-off-by: Ido Yariv <ido@wizery.com>