wl12xx: add set_ssid_filter to scan command
The ssid  filtering in the fw doesn't handle well p2p
scans, as the "DIRECT-" is treated as wildcard, while
the filter looks for exact match.

A new set_ssid_filter field was added to the scan command
in order to disable the ssid filtering in specific cases.

Enable ssid filtering only for targeted (ssid-specifc) scans.