mac80211: fix race condition between assoc_done and first EAP packet
When associating to an AP, the station might miss the first EAP
packet that the AP sends due to a race condition between the association
success procedure and the rx flow in mac80211.
In such cases, the packet might fall in ieee80211_rx_h_check due to
the fact that the relevant rx->sta wasn't allocated yet.
Allocation of the relevant station info struct before actually
sending the association request and setting it with a new
WLAN_STA_PRE_ASSOC pseudo state solve this problem.
The station will except only EAP packets from the AP while it
is in the pre-association state.

Signed-off-by: Guy Eilam <guy@wizery.com>