mac80211: disassociate before authentication (EXPERIMENTAL)
On roaming, we should disconnect from the old AP (?) before
connecting to the new one. do it by disassociating before
trying to authenticate.

This patch helps devices that encounter problems with offchannel
tx (as it eliminates the need to send auth req on one channel
while being connected on a different channel)